requirements.txt
llama-cpp-python==0.3.1
Flask==3.0.3
Flask-Cors==3.0.10
pymupdf==1.24.11
easyocr==1.7.2